put the things in the right places
